requirements.txt
PyYAML>=6.0
Faker>=20.0.0
